chore(broker): move source files into brokers/ and http/ subtrees
Separates the flat src/ layout into two subtrees: src/brokers/ holds the
data layer (with stream/ and persistence/ for shared infrastructure), and
src/http/ holds the HTTP transport layer (middleware/, routes/).

Architectural changes included:
- completion-chunk-broker.ts renamed to chunks-broker.ts
- main.ts renamed to index.ts; package.json entry points updated
- spanMatchesSessionId moved into trace-broker.ts, removing the broker→route
  circular import that existed before
- CRD-shadow types (QueryPhase, EventReasons, SessionEventData, etc.) inlined
  into sessions-broker.ts; StreamError inlined into http/routes/stream.ts;
  dead types (Message, StoredMessage, Add/MessagesRequest/Response) dropped;
  src/types.ts deleted
- @ark-broker/ path alias used for all cross-tree imports; jest.config.js
  moduleNameMapper updated to strip .js extensions from alias paths
