+++ services/ark-broker/ark-broker/jest.config.js
@@ -25,6 +25,7 @@ export default {
     },
   },
   moduleNameMapper: {
+    '^@ark-broker/(.*)\\.js$': '<rootDir>/src/$1',
     '^@ark-broker/(.*)$': '<rootDir>/src/$1',
     '^(\\.{1,2}/.*)\\.js$': '$1',
   },
