+++ services/ark-broker/ark-broker/package.json
@@ -2,12 +2,12 @@
   "name": "ark-broker",
   "version": "0.1.63",
   "description": "Ark broker for memory, telemetry streaming, and observability",
-  "main": "dist/main.js",
+  "main": "dist/index.js",
   "type": "module",
   "scripts": {
     "build": "tsc && tsc-alias",
-    "dev": "NODE_ENV=development tsx watch --clear-screen=false src/main.ts",
-    "start": "node dist/main.js",
+    "dev": "NODE_ENV=development tsx watch --clear-screen=false src/index.ts",
+    "start": "node dist/index.js",
     "test": "jest",
     "test:watch": "jest --watch",
     "test:coverage": "jest --coverage",
