+++ services/ark-broker/ark-broker/src/brokers/event-broker.ts
@@ -1,6 +1,6 @@
-import {BrokerItem} from './broker-item.js';
-import {BrokerItemStream} from './broker-item-stream.js';
-import type {Logger} from './logging/logger.js';
+import {BrokerItem} from './stream/broker-item.js';
+import {BrokerItemStream} from './stream/broker-item-stream.js';
+import type {Logger} from '../logging/logger.js';
 import {PaginatedList, PaginationParams} from './pagination.js';
 
 /** Event data from Ark controller operations */
