+++ services/ark-broker/ark-broker/src/http/middleware/error-handler.ts
